% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_SFA_Est.R
\name{il_sfa}
\alias{il_sfa}
\title{Integrated Likelihood estimation function}
\usage{
il_sfa(
  X,
  y,
  distr = "Exp",
  het = TRUE,
  z,
  group,
  nq = 25,
  niter = 10,
  init = NULL,
  useHess = TRUE,
  trace = 0,
  grtol = 10^-6,
  grad = "central"
)
}
\arguments{
\item{X}{the matrix of independent variables}

\item{y}{the response variable vector}

\item{distr}{the distributional form of the efficiency related error terms (default is Exp)}

\item{het}{a logical indicating the}

\item{z}{the matrix of variables influencing the efficiency}

\item{group}{the vector of grouping variable}

\item{nq}{the number of quadrature points used in the numerical integration routine}

\item{niter}{the number of iterations used in the numerical integration routine}

\item{init}{the vector of initial values for likelihood parameters}

\item{useHess}{a logical: if TRUE the likelihood maximization is initialized with the numerical Hessian otherwise the Hessian is initialized as a unit diagonal}

\item{trace}{trace parameter used in ucminf maximization algorithm}

\item{grtol}{grtol parameter used in ucminf maximization algorithm (gradient tollerance)}

\item{grad}{grad parameter used in ucminf maximization algorithm}
}
\value{
a list including the point estimates, the numerical hessian, the standard errors of the estimates and the firm specific effects (alpha_i) and the estimated efficiency related terms (u_it).
}
\description{
Adopts the IL approach to the estimation of Stochastic Frontier Model via marginal likelihood.
}
\examples{
### Reproduce one single dataset with the setting used in Table 1, (a) lambda = 1
### Belotti TFE - N=100, T=5
N <- 100
T <- 5
para <- c(1,-1.5,1,0.25)
dat <- DGP(N, T, list(beta=para[1], sigmav=para[4], gamma=para[2:3]), model = "Exp",
           seed.val = 22, nsim = 1)
mydat <- list(y = dat$daty[1,], x = data.frame(dat$x), g = dat$g, z = dat$datz)
mle <- il_sfa(X = mydat$x, y = mydat$y, distr = "Exp", het = TRUE, z = mydat$z,
              group = mydat$g, useHess = TRUE)
mle

}
