% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_SFA_Sim.R
\name{DGP}
\alias{DGP}
\title{The Data Generating Process used in the simulation studies.}
\usage{
DGP(n, m, param, model = "HN", seed.val = 0, nsim = 1000, zvar = FALSE)
}
\arguments{
\item{n}{the cross-sectional dimension.}

\item{m}{the time dimension.}

\item{param}{the vector of model parameters (the frontier regression parameters \eqn{\beta},
the idiosyncratic error variance \eqn{\sigma_v},
the vector of the parameters describing the efficiency behavior \eqn{\gamma}). \eqn{\gamma} depends on \eqn{u_{it}} distribution choice. 
In the half-normal case, it corresponds to \eqn{\sigma_u}. In the exponential case, it collects \eqn{\gamma_0} and \eqn{\gamma_1} defined in the formula above.
In the gamma case, it corresponds to .}

\item{model}{the kind of distribution considered to generate the efficiency-related terms.}

\item{seed.val}{the seed value used to make the simulations reproducible.}

\item{nsim}{an integer indicating the number of simulated datasets.}

\item{zvar}{a logical identifying the case in which the efficiency is time variant.}
}
\value{
a list object including:

-- the simulated y vectors (a matrix daty of dimension \eqn{nsim \times n\cdot m}),

-- the simulated of efficiency-related terms (a matrix datu of dimension \eqn{nsim \times n\cdot m}),

-- the simulated of independent variables (a matrix \eqn{x} of dimension \eqn{n\cdot m \times p}),

-- the simulated of grouping identifiers (a vector \eqn{g} of length \eqn{n\cdot m}), and

-- the simulated values of the variable influencing the efficiency level (a vector \eqn{datz} of length \eqn{n\cdot m}).
}
\description{
This code allows to reproduce the results in Table 1 panel (a)
Belotti and Ilardi (2018) "Consistent inference in fixed-effects
stochastic frontier models" - The sample size is fixed to N=100 and T=5.
The example is limited to 10 replications to make the code
fairly executable. Changing the nsim parameter, one can consider the
original simulation with 1000 replications. set.seed = 22 is used to generate the
data analyzed with Stata and R and compared in Bellio and Grassetti (2023).
In the cited papers, the simulations use a number of replications equal to 1000.
The available distributions for the efficiency term are homoschedastic half-normal ("HN"), 
homoschedastic and heteroschedastic Exponential ("Exp"),
and homoschedastic Gamma ("Gamma").
}
\details{
The DGP is defined as
\deqn{y_{it} = \alpha_i + x_{it}^T \beta - u_{it} + v_{it}}
where the distribution of \eqn{u_{it}} can be Half-Normal, 
Gamma, or Exponential. In this last
case, the mean of the Exponential distribution can be defined as
\deqn{\sigma_{uit} = \exp\left(\gamma_0 + \gamma_1 z_{it}\right).}
This last definition also brings to time-constant cases when
\eqn{z_{it}} is considered constant over time.
}
\examples{

N <- 100
T <- 5
para <- c(1,-1.5,1,0.25)
nsimul <- 10
dat <- DGP(N, T, list(beta = para[1], sigmav = para[4], gamma = para[2:3]), model = "Exp",
           seed.val = 22, nsim = nsimul)
res <- matrix(0, nrow = nsimul, ncol=4)
for(i in 1:nrow(res)){
  print(i)
  mydat <- list(y = dat$daty[i,], x = data.frame(dat$x), g = dat$g, z = dat$datz)
  mle <- il_sfa(X = mydat$x, y = mydat$y, distr = "Exp", het = TRUE, z = mydat$z,
                group = mydat$g, useHess = TRUE)
  res[i,] <- mle$par
  }
# Compute the MSE of the estimates
for(i in 1:4) print(MSE(res[,i], para[i]))
# Compare them with the estimated values included in the
# package data (res_N100_Stata.rda)
data(res_N100_Stata)
for(i in 1:4) print(MSE(res_N100_Stata[1:nsimul,i], para[i]))

}
\references{
Belotti, F., & Ilardi, G. (2018). Consistent inference in
fixed-effects stochastic frontier models. Journal of Econometrics, 202(2), 161--177.

Bellio, R., & Grassetti, L. (2023). Efficient estimation of true
fixed-effects stochastic frontier models. Submitted
}
