% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL_SFA_Est.R
\name{il_sfa}
\alias{il_sfa}
\title{True Fixed Effects Stochastic Frontier Model estimation via Integrated Likelihood.}
\usage{
il_sfa(
  X,
  y,
  distr = "Exp",
  het = TRUE,
  z,
  group,
  nq = 25,
  niter = 10,
  init = NULL,
  useHess = TRUE,
  trace = 0,
  grtol = 10^-6,
  grad = "central"
)
}
\arguments{
\item{X}{the matrix of independent variables.}

\item{y}{the vector of the response variable.}

\item{distr}{the distributional form of the efficiency-related error terms (default is Exp).}

\item{het}{a logical indicating the case of heteroscedastic efficiency related error
terms (default is TRUE, but it works for the Exponential case only).}

\item{z}{the vector of values of the variable influencing the efficiency (for the Exponential case only).}

\item{group}{the vector of the grouping variable defining the panel.}

\item{nq}{the number of quadrature points used in the numerical integration routine (default is 25).}

\item{niter}{the number of iterations used in the numerical integration routine for locating the mode (default is 10).}

\item{init}{the optional vector of initial values for the maximization of integrated likelihood (default is NULL).}

\item{useHess}{a logical: if TRUE, the likelihood maximization is initialized with the numerical
Hessian otherwise, the Hessian is initialized as a unit diagonal.}

\item{trace}{trace parameter used in ucminf function (if trace is positive, then
detailed tracing information is printed for each iteration).}

\item{grtol}{grtol parameter used in ucminf function (the algorithm stops when
the largest absolute value of the gradient is less than grtol).}

\item{grad}{grad parameter used in ucminf function (controls the type
of finite difference approximation to be used for the gradient).}
}
\value{
A list including

-- the parameters point estimates - \eqn{par},

-- the inverse of the numerical hessian - \eqn{invhes},

-- the standard errors of the estimated parameters - \eqn{se}

-- the firm specific effects - \eqn{alphai}, and

-- the estimated efficiency related terms -  - \eqn{uit}.
}
\description{
This function adopts the IL approach to the estimation of some kind of SF models
(see Bellio and Grassetti, 2023) such
as those based on N-HN and N-Exponential error terms convolutions.
The function adopts a consistent and computationally efficient approach to the
model estimation.
The model specification is as
\deqn{y_{ij} = \alpha_i + x_{it}^T\beta - u_{it} + v{it},}
where the efficiency-related error term can be estimated as \eqn{u_{it} \sim HN(0, \sigma_u^2)},
\eqn{u_{it} \sim Exp(\sigma_u^2)}, or \eqn{u_{it} \sim Exp(\sigma_{ui}^2)} and
\eqn{\sigma_{ui} = e^{\gamma_0+\gamma_1z_i}}.
The idiosyncratic error term is considered Gaussian and homoscedastic.
}
\examples{
# Generate one single dataset with the setting used in Table 1, (a)
# Belotti and Ilardi (2018) "Consistent inference in fixed-effects
# stochastic frontier models" - N=100, T=5.

N <- 100
T <- 5
para <- c(1,-1.5,1,0.25)
dat <- DGP(N, T, list(beta=para[1], sigmav=para[4], gamma=para[2:3]), model = "Exp",
           seed.val = 22, nsim = 1)
mydat <- list(y = dat$daty[1,], x = data.frame(dat$x), g = dat$g, z = dat$datz)
mle <- il_sfa(X = mydat$x, y = mydat$y, distr = "Exp", het = TRUE, z = mydat$z,
              group = mydat$g, useHess = TRUE)
mle

}
\references{
Belotti, F., & Ilardi, G. (2018). Consistent inference in
fixed-effects stochastic frontier models. Journal of Econometrics, 202(2), 161--177.

Bellio, R., & Grassetti, L. (2023). Efficient estimation of true
fixed-effects stochastic frontier models. Submitted
}
