% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res_N100_Stata}
\alias{res_N100_Stata}
\title{Results from the Stata estimation function using PDE method}
\format{
A data frame with 1000 rows and 4 variables (the model parameters)
 \describe{
     \item{b1}{The slope parameter in the frontier function}
     \item{b2}{The heteroschedasticity parameter gamma_0}
     \item{b3}{The heteroschedasticity parameter gamma_1}
     \item{b4}{The idiosyncratic error variance}
     }
}
\source{
{Stata estimation results obtained considering the DGP described in the paper}
}
\usage{
res_N100_Stata
}
\description{
Contains the estimation results of the 1000 replications of the simulation experiment also developed in this package
}
\examples{
data(res_N100_Stata)
}
\keyword{datasets}
